Update: Substring Error Fix
If an application implements some kind of std output, then the test adapter was counting that as valid output (skipping the IsNullOrEmpty check) and trying to find the doctest start/end search strings even when the search string isn't included in the .exe output. So for now, checking if the search string is in the ouput at all before trying to access the doctest output.